upd prettier, asset modify & approve
