+++ drizzle.config.ts
@@ -1,10 +1,10 @@
 import type { Config } from "drizzle-kit"
 
 export default {
-    out: "./src/db/migrations",
-    schema: "./src/db/schema.ts",
-    driver: "turso",
-    breakpoints: true,
-    strict: true,
-    verbose: true,
+	out: "./src/db/migrations",
+	schema: "./src/db/schema.ts",
+	driver: "turso",
+	breakpoints: true,
+	strict: true,
+	verbose: true,
 } satisfies Config
