+++ package.json
@@ -1,39 +1,39 @@
 {
-    "name": "wanderer-moe-api",
-    "version": "1.0.1b",
-    "scripts": {
-        "prettier:fmt": "prettier --write .",
-        "dev": "wrangler dev --remote",
-        "publish": "wrangler publish --minify",
-        "lint": "eslint . --ext .ts",
-        "prettier:check": "prettier --check .",
-        "typecheck": "tsc --noEmit",
-        "drizzle:generate": "drizzle-kit generate:sqlite",
-        "drizzle:push": "drizzle-kit push:sqlite"
-    },
-    "devDependencies": {
-        "@cloudflare/workers-types": "^4.20230821.0",
-        "@types/node": "^20.5.7",
-        "dotenv": "^16.3.1",
-        "drizzle-kit": "^0.19.13",
-        "eslint": "^8.48.0",
-        "eslint-config-google": "^0.14.0",
-        "eslint-plugin-json": "^3.1.0",
-        "husky": "^8.0.3",
-        "tsx": "^3.12.7",
-        "typescript": "^5.2.2",
-        "wrangler": "3.6.0"
-    },
-    "private": true,
-    "dependencies": {
-        "@libsql/client": "^0.3.2",
-        "@lucia-auth/adapter-sqlite": "^2.0.0",
-        "@typescript-eslint/eslint-plugin": "^6.5.0",
-        "drizzle-orm": "^0.28.5",
-        "hono": "^3.5.6",
-        "lucia": "^2.4.0",
-        "mysql2": "^3.6.0",
-        "prettier": "^3.0.3",
-        "resend": "^1.0.0"
-    }
+	"name": "wanderer-moe-api",
+	"version": "1.0.1b",
+	"scripts": {
+		"prettier:fmt": "prettier --write .",
+		"dev": "wrangler dev --remote",
+		"publish": "wrangler publish --minify",
+		"lint": "eslint . --ext .ts",
+		"prettier:check": "prettier --check .",
+		"typecheck": "tsc --noEmit",
+		"drizzle:generate": "drizzle-kit generate:sqlite",
+		"drizzle:push": "drizzle-kit push:sqlite"
+	},
+	"devDependencies": {
+		"@cloudflare/workers-types": "^4.20230821.0",
+		"@types/node": "^20.5.7",
+		"dotenv": "^16.3.1",
+		"drizzle-kit": "^0.19.13",
+		"eslint": "^8.48.0",
+		"eslint-config-google": "^0.14.0",
+		"eslint-plugin-json": "^3.1.0",
+		"husky": "^8.0.3",
+		"tsx": "^3.12.7",
+		"typescript": "^5.2.2",
+		"wrangler": "3.6.0"
+	},
+	"private": true,
+	"dependencies": {
+		"@libsql/client": "^0.3.2",
+		"@lucia-auth/adapter-sqlite": "^2.0.0",
+		"@typescript-eslint/eslint-plugin": "^6.5.0",
+		"drizzle-orm": "^0.28.5",
+		"hono": "^3.5.6",
+		"lucia": "^2.4.0",
+		"mysql2": "^3.6.0",
+		"prettier": "^3.0.3",
+		"resend": "^1.0.0"
+	}
 }
