+++ src/index.ts
@@ -11,15 +11,15 @@ import { Bindings } from "@/worker-configuration"
 const app = new Hono<{ Bindings: Bindings }>()
 
 app.get("/status", (c) => {
-    c.status(200)
-    return c.json({
-        status: "ok",
-        runtime: getRuntimeKey(),
-    })
+	c.status(200)
+	return c.json({
+		status: "ok",
+		runtime: getRuntimeKey(),
+	})
 })
 app.get("/", (c) => {
-    c.status(200)
-    return c.json({ success: "true", status: "ok", routes: app.routes })
+	c.status(200)
+	return c.json({ success: "true", status: "ok", routes: app.routes })
 })
 app.route("/v2/asset", assetRoute)
 app.route("/v2/discord", discordRoute)
@@ -28,8 +28,8 @@ app.route("/v2/search", searchRoute)
 app.route("/v2/games", gamesRoute)
 app.route("/v2/auth", authRoute)
 app.all("*", (c) => {
-    c.status(404)
-    return c.json({ status: "not found" })
+	c.status(404)
+	return c.json({ status: "not found" })
 })
 
 // https://hono.dev/api/hono#showroutes
