+++ src/v2/db/drizzle.ts
@@ -1,23 +1,23 @@
 export const tableNames = {
-    assets: "assets",
-    authKey: "authKey",
-    authSession: "authSession",
-    authUser: "authUser",
-    emailVerificationTokens: "emailVerificationToken",
-    follower: "follower",
-    following: "following",
-    gameAssetCategories: "gameAssetCategories",
-    games: "games",
-    assetTags: "assetTags",
-    emailVerificationToken: "emailVerificationToken",
-    passwordResetToken: "passwordResetToken",
-    assetCategories: "assetCategories",
-    passwordResetTokens: "passwordResetToken",
-    savedOcGenerators: "savedOcGenerators",
-    savedAssets: "savedAssets",
-    assetCollection: "assetCollection",
-    assetCollectionAsset: "assetCollectionAsset",
-    socialsConnection: "socialsConnection",
+	assets: "assets",
+	authKey: "authKey",
+	authSession: "authSession",
+	authUser: "authUser",
+	emailVerificationTokens: "emailVerificationToken",
+	follower: "follower",
+	following: "following",
+	gameAssetCategories: "gameAssetCategories",
+	games: "games",
+	assetTags: "assetTags",
+	emailVerificationToken: "emailVerificationToken",
+	passwordResetToken: "passwordResetToken",
+	assetCategories: "assetCategories",
+	passwordResetTokens: "passwordResetToken",
+	savedOcGenerators: "savedOcGenerators",
+	savedAssets: "savedAssets",
+	assetCollection: "assetCollection",
+	assetCollectionAsset: "assetCollectionAsset",
+	socialsConnection: "socialsConnection",
 }
 
 export * as schema from "@/v2/db/schema"
