+++ src/v2/db/turso.ts
@@ -5,18 +5,18 @@ import { createClient } from "@libsql/client/web" // because we're in a worker
 
 // wrapper for turso & drizzle
 export function getConnection(env: Env) {
-    const turso = createClient({
-        url: env.TURSO_DATABASE_URL,
-        authToken: env.TURSO_DATABASE_AUTH_TOKEN,
-    })
+	const turso = createClient({
+		url: env.TURSO_DATABASE_URL,
+		authToken: env.TURSO_DATABASE_AUTH_TOKEN,
+	})
 
-    const drizzle = drizzleORM(turso, {
-        schema,
-        logger: true,
-    })
+	const drizzle = drizzleORM(turso, {
+		schema,
+		logger: true,
+	})
 
-    return {
-        drizzle,
-        turso,
-    }
+	return {
+		drizzle,
+		turso,
+	}
 }
