+++ src/v2/lib/auth/roleFlags.ts
@@ -2,41 +2,41 @@
 
 // permission based roles
 export const roleFlags = {
-    USER: 1 << 0,
-    CONTRIBUTOR: 1 << 1,
-    TRANSLATOR: 1 << 2,
-    STAFF: 1 << 3,
-    DEVELOPER: 1 << 4,
-    CREATOR: 1 << 5,
+	USER: 1 << 0,
+	CONTRIBUTOR: 1 << 1,
+	TRANSLATOR: 1 << 2,
+	STAFF: 1 << 3,
+	DEVELOPER: 1 << 4,
+	CREATOR: 1 << 5,
 }
 
 export const roleFlagsToArray = (roleFlagsInt: number): string[] => {
-    const roles: string[] = []
+	const roles: string[] = []
 
-    for (const [role, flag] of Object.entries(roleFlags)) {
-        if (roleFlagsInt & flag) roles.push(role)
-    }
+	for (const [role, flag] of Object.entries(roleFlags)) {
+		if (roleFlagsInt & flag) roles.push(role)
+	}
 
-    return roles
+	return roles
 }
 
 // self assignable roles
 export const SelfAssignableRoleFlags = {
-    CONTENT_CREATOR: 1 << 0,
-    ARTIST: 1 << 1,
-    WRITER: 1 << 2,
-    DEVELOPER: 1 << 3,
-    DESIGNER: 1 << 4,
+	CONTENT_CREATOR: 1 << 0,
+	ARTIST: 1 << 1,
+	WRITER: 1 << 2,
+	DEVELOPER: 1 << 3,
+	DESIGNER: 1 << 4,
 }
 
 export const SelfAssignableRoleFlagsToArray = (
-    roleFlagsInt: number
+	roleFlagsInt: number
 ): string[] => {
-    const roles: string[] = []
+	const roles: string[] = []
 
-    for (const [role, flag] of Object.entries(SelfAssignableRoleFlags)) {
-        if (roleFlagsInt & flag) roles.push(role)
-    }
+	for (const [role, flag] of Object.entries(SelfAssignableRoleFlags)) {
+		if (roleFlagsInt & flag) roles.push(role)
+	}
 
-    return roles
+	return roles
 }
