+++ src/v2/lib/discord.ts
@@ -1,12 +1,12 @@
 export const roles: { [key: string]: string } = {
-    "982387185259012096": "Project Lead",
-    "1112182244572938310": "Developer",
-    "1038892176479895584": "Admin",
-    "1000916582538674249": "Senior Moderator",
-    "983817984923562034": "Moderator",
-    "1088105796908355584": "Translator",
-    "1005805438031364129": "Contributor",
-    "983883539772751912": "Server Booster",
+	"982387185259012096": "Project Lead",
+	"1112182244572938310": "Developer",
+	"1038892176479895584": "Admin",
+	"1000916582538674249": "Senior Moderator",
+	"983817984923562034": "Moderator",
+	"1088105796908355584": "Translator",
+	"1005805438031364129": "Contributor",
+	"983883539772751912": "Server Booster",
 }
 
 export const guildId = "982385887000272956"
