+++ src/v2/lib/helpers/responses/notFoundResponse.ts
@@ -1,12 +1,12 @@
 // helper function to create a 404 Not Found response
 export function createNotFoundResponse(c, errorMessage, responseHeaders) {
-    return c.json(
-        {
-            success: false,
-            status: "error",
-            error: errorMessage,
-        },
-        404,
-        responseHeaders
-    )
+	return c.json(
+		{
+			success: false,
+			status: "error",
+			error: errorMessage,
+		},
+		404,
+		responseHeaders
+	)
 }
