+++ src/v2/lib/listBucket.ts
@@ -1,3 +1,3 @@
 export const listBucket = async (bucket, options) => {
-    return await bucket.list(options)
+	return await bucket.list(options)
 }
