+++ src/v2/lib/responseHeaders.ts
@@ -1,5 +1,5 @@
 export const responseHeaders: Record<string, string> = {
-    "X-Content-Type-Options": "nosniff",
-    "Referrer-Policy": "strict-origin-when-cross-origin",
-    "content-type": "application/json;charset=UTF-8",
+	"X-Content-Type-Options": "nosniff",
+	"Referrer-Policy": "strict-origin-when-cross-origin",
+	"content-type": "application/json;charset=UTF-8",
 }
