+++ src/v2/routes/asset/assetRoute.ts
@@ -6,12 +6,12 @@ import { Bindings } from "@/worker-configuration"
 const assetRoute = new Hono<{ Bindings: Bindings }>()
 
 assetRoute.get("/:id", async (c) => {
-    return getAssetFromId(c)
+	return getAssetFromId(c)
 })
 
 // setting both of these to id returns "duplicate param name" error, will fix later
 assetRoute.get("/download/:assetId", async (c) => {
-    return downloadAsset(c)
+	return downloadAsset(c)
 })
 
 export default assetRoute
