+++ src/v2/routes/auth/authRoute.ts
@@ -8,63 +8,73 @@ import { uploadProfileImage } from "./user-attributes/self-upload/uploadAvatar"
 import { uploadBannerImage } from "./user-attributes/self-upload/uploadBanner"
 import { saveOCGeneratorResponse } from "./oc-generators/saveOCGeneratorResponse"
 import { updateUserAttributes } from "./user-attributes/updateUserAttributes"
-import { uploadAsset } from "./user-attributes/self-upload/uploadAsset"
+import { uploadAsset } from "./assets/uploadAsset"
 import { Bindings } from "@/worker-configuration"
+import { modifyAssetData } from "./assets/modifyAsset"
+import { approveAsset } from "./assets/approveAsset"
 import { viewOCGeneratorResponses } from "./oc-generators/viewOCGeneratorResponses"
 import { deleteOCGeneratorResponse } from "./oc-generators/deleteOCGeneratorResponse"
 
 const authRoute = new Hono<{ Bindings: Bindings }>()
 
 authRoute.use(
-    "*",
-    cors({
-        credentials: true,
-        origin: ["https://next.wanderer.moe"],
-    })
+	"*",
+	cors({
+		credentials: true,
+		origin: ["https://next.wanderer.moe"],
+	})
 )
 
 authRoute.post("/login", async (c) => {
-    return login(c)
+	return login(c)
 })
 
 authRoute.post("/update/attributes", async (c) => {
-    return updateUserAttributes(c)
+	return updateUserAttributes(c)
 })
 
 authRoute.post("/upload/asset", async (c) => {
-    return uploadAsset(c)
+	return uploadAsset(c)
 })
 
 authRoute.post("/upload/avatar", async (c) => {
-    return uploadProfileImage(c)
+	return uploadProfileImage(c)
+})
+
+authRoute.post("/approve/asset/:assetIdToApprove", async (c) => {
+	return approveAsset(c)
+})
+
+authRoute.post("/modify/asset/:assetIdToModify", async (c) => {
+	return modifyAssetData(c)
 })
 
 authRoute.post("/upload/banner", async (c) => {
-    return uploadBannerImage(c)
+	return uploadBannerImage(c)
 })
 
 authRoute.post("/signup", async (c) => {
-    return signup(c)
+	return signup(c)
 })
 
 authRoute.post("/oc-generator/save", async (c) => {
-    return saveOCGeneratorResponse(c)
+	return saveOCGeneratorResponse(c)
 })
 
 authRoute.get("/oc-generator/view/all", async (c) => {
-    return viewOCGeneratorResponses(c)
+	return viewOCGeneratorResponses(c)
 })
 
 authRoute.post("/oc-generator/delete", async (c) => {
-    return deleteOCGeneratorResponse(c)
+	return deleteOCGeneratorResponse(c)
 })
 
 authRoute.get("/validate", async (c) => {
-    return validate(c)
+	return validate(c)
 })
 
 authRoute.post("/logout", async (c) => {
-    return logout(c)
+	return logout(c)
 })
 
 export default authRoute
