+++ src/v2/routes/discord/discordRoute.ts
@@ -5,7 +5,7 @@ import { Bindings } from "@/worker-configuration"
 const discordRoute = new Hono<{ Bindings: Bindings }>()
 
 discordRoute.get("/contributors", async (c) => {
-    return contributors(c)
+	return contributors(c)
 })
 
 export default discordRoute
