+++ src/v2/routes/games/gamesRoute.ts
@@ -5,7 +5,7 @@ import { Bindings } from "@/worker-configuration"
 const gamesRoute = new Hono<{ Bindings: Bindings }>()
 
 gamesRoute.get("/all", async (c) => {
-    return getAllGames(c)
+	return getAllGames(c)
 })
 
 export default gamesRoute
