+++ src/v2/routes/oc-generators/ocGeneratorRoutes.ts
@@ -6,11 +6,11 @@ import { Bindings } from "@/worker-configuration"
 const ocGeneratorRoute = new Hono<{ Bindings: Bindings }>()
 
 ocGeneratorRoute.get("/", async (c) => {
-    return getGenerators(c)
+	return getGenerators(c)
 })
 
 ocGeneratorRoute.get("/:gameName", async (c) => {
-    return getGeneratorFromName(c)
+	return getGeneratorFromName(c)
 })
 
 export default ocGeneratorRoute
