+++ src/v2/routes/search/searchRoute.ts
@@ -9,23 +9,23 @@ import { cors } from "hono/cors"
 const searchRoute = new Hono<{ Bindings: Bindings }>()
 
 searchRoute.get("/all", async (c) => {
-    return searchAll(c)
+	return searchAll(c)
 })
 
 authRoute.use(
-    "/all",
-    cors({
-        credentials: true,
-        origin: ["https://next.wanderer.moe"],
-    })
+	"/all",
+	cors({
+		credentials: true,
+		origin: ["https://next.wanderer.moe"],
+	})
 )
 
 searchRoute.get("/users/user/:username", async (c) => {
-    return getUserByUsername(c)
+	return getUserByUsername(c)
 })
 
 searchRoute.get("/users/query/:query", async (c) => {
-    return getUsersBySearch(c)
+	return getUsersBySearch(c)
 })
 
 export default searchRoute
