+++ src/v2/routes/search/user/getUserByUsername.ts
@@ -4,41 +4,41 @@ import { createNotFoundResponse } from "@/v2/lib/helpers/responses/notFoundRespo
 import type { APIContext as Context } from "@/worker-configuration"
 
 export async function getUserByUsername(c: Context): Promise<Response> {
-    const { username } = c.req.param()
-    const cacheKey = new Request(c.req.url.toString(), c.req)
-    const cache = caches.default
-    let response = await cache.match(cacheKey)
+	const { username } = c.req.param()
+	const cacheKey = new Request(c.req.url.toString(), c.req)
+	const cache = caches.default
+	let response = await cache.match(cacheKey)
 
-    if (response) return response
+	if (response) return response
 
-    const drizzle = await getConnection(c.env).drizzle
+	const drizzle = await getConnection(c.env).drizzle
 
-    const user = await drizzle.query.users.findFirst({
-        where: (users, { eq }) => eq(user.username, username),
-    })
+	const user = await drizzle.query.users.findFirst({
+		where: (users, { eq }) => eq(user.username, username),
+	})
 
-    if (!user) {
-        response = createNotFoundResponse(c, "User not found", responseHeaders)
-        await cache.put(cacheKey, response.clone())
-        return response
-    }
+	if (!user) {
+		response = createNotFoundResponse(c, "User not found", responseHeaders)
+		await cache.put(cacheKey, response.clone())
+		return response
+	}
 
-    // removing email-related fields
-    user.email = undefined
-    user.emailVerified = undefined
+	// removing email-related fields
+	user.email = undefined
+	user.emailVerified = undefined
 
-    response = c.json(
-        {
-            success: true,
-            status: "ok",
-            user,
-        },
-        200,
-        responseHeaders
-    )
+	response = c.json(
+		{
+			success: true,
+			status: "ok",
+			user,
+		},
+		200,
+		responseHeaders
+	)
 
-    response.headers.set("Cache-Control", "s-maxage=300")
-    await cache.put(cacheKey, response.clone())
+	response.headers.set("Cache-Control", "s-maxage=300")
+	await cache.put(cacheKey, response.clone())
 
-    return response
+	return response
 }
