+++ src/v2/routes/search/user/getUsersBySearch.ts
@@ -6,37 +6,37 @@ import { users } from "@/v2/db/schema"
 import type { APIContext as Context } from "@/worker-configuration"
 
 export async function getUsersBySearch(c: Context): Promise<Response> {
-    const cacheKey = new Request(c.req.url.toString(), c.req)
-    const cache = caches.default
-    let response = await cache.match(cacheKey)
-    if (response) return response
+	const cacheKey = new Request(c.req.url.toString(), c.req)
+	const cache = caches.default
+	let response = await cache.match(cacheKey)
+	if (response) return response
 
-    const { query } = c.req.param()
-    const drizzle = await getConnection(c.env).drizzle
+	const { query } = c.req.param()
+	const drizzle = await getConnection(c.env).drizzle
 
-    const userList = await drizzle
-        .select()
-        .from(users)
-        .where(like(users.username, `%${query}%`))
-        .execute()
+	const userList = await drizzle
+		.select()
+		.from(users)
+		.where(like(users.username, `%${query}%`))
+		.execute()
 
-    if (!userList) {
-        return createNotFoundResponse(c, "Users not found", responseHeaders)
-    }
+	if (!userList) {
+		return createNotFoundResponse(c, "Users not found", responseHeaders)
+	}
 
-    response = c.json(
-        {
-            success: true,
-            status: "ok",
-            query,
-            results: userList,
-        },
-        200,
-        responseHeaders
-    )
+	response = c.json(
+		{
+			success: true,
+			status: "ok",
+			query,
+			results: userList,
+		},
+		200,
+		responseHeaders
+	)
 
-    response.headers.set("Cache-Control", "s-maxage=60")
-    await cache.put(cacheKey, response.clone())
+	response.headers.set("Cache-Control", "s-maxage=60")
+	await cache.put(cacheKey, response.clone())
 
-    return response
+	return response
 }
